Rearrange and cleanup AST classes
Rename the modules to have more meaningful names
