+++ lib/ast/index.ts
@@ -34,11 +34,14 @@ export * from './base';
 export * from './function_def';
 export * from './class_def';
 export * from './values';
+export * from './invocation';
+export * from './boolean_expression';
 export * from './expression';
-export * from './expression2';
-export * from './primitive';
+export * from './permissions';
+export * from './legacy';
+export * from './statement';
 export * from './program';
-export * from './bookkeeping';
+export * from './control_commands';
 export * from './dialogues';
 export { default as NodeVisitor } from './visitor';
-export { AbstractSlot } from './slots';
+export * from './slots';
